+++ management/src/commands/add-identity-to-user.ts
@@ -1,8 +1,14 @@
 import gql from "graphql-tag";
 
-import { GraphqlErrors, OutputDataNullError } from "../errors";
-import { fetchManagement } from "../fetch-management";
 import {
+  GraphqlErrors,
+  InvalidValidationCodeError,
+  OutputDataNullError,
+} from "../errors";
+import { fetchManagement } from "../fetch-management";
+import { checkVerificationCode } from "../queries/check-verification-code";
+import {
+  CheckVerificationCodeStatus,
   Identity,
   IdentityCommandInput,
   ManagementCredentials,
@@ -26,10 +32,36 @@ const ADD_IDENTITY_TO_USER = gql`
   }
 `;
 
-export async function addIdentityToUser(
+async function addIdentityToUser(
   managementCredentials: ManagementCredentials,
+  validationCode: string,
+  eventIds: string[],
   { userId, identityType, identityValue }: IdentityCommandInput,
 ): Promise<Identity> {
+  let validationStatus: CheckVerificationCodeStatus.VALID | undefined;
+
+  for await (const eventId of eventIds.reverse()) {
+    if (validationStatus === CheckVerificationCodeStatus.VALID) {
+      break;
+    }
+
+    const { status: verifiedResult } = await checkVerificationCode(
+      managementCredentials,
+      {
+        code: validationCode,
+        eventId,
+      },
+    );
+
+    if (verifiedResult === CheckVerificationCodeStatus.VALID) {
+      validationStatus = verifiedResult;
+    }
+  }
+
+  if (!validationStatus) {
+    throw new InvalidValidationCodeError();
+  }
+
   const operation = {
     query: ADD_IDENTITY_TO_USER,
     variables: { userId, type: identityType, value: identityValue },
@@ -49,3 +81,5 @@ export async function addIdentityToUser(
 
   return data.addIdentityToUser;
 }
+
+export { addIdentityToUser };